Fixed audio input selection

+++ libbeat/soundbuffer.cpp
@@ -20,51 +20,56 @@
 namespace libbeat
 {
 
-SoundBuffer::SoundBuffer(uint16_t size)
+SoundBuffer::SoundBuffer(int size)
 {
-    this->m_size=size;
+    this->m_size = size;
     m_Buffer.resize(size);
 }
 
-int16_t SoundBuffer::average()
+int SoundBuffer::average()
 {
-    int32_t sum=0;
-    for(uint16_t i=0;i<m_size;i++)
-    {
-        sum+=m_Buffer[i];
+    int sum = 0;
+    for(int i = 0; i < m_size; i++){
+        sum += m_Buffer[i];
     }
-    sum/=m_size;
-    return (int16_t)sum;
+    sum /= m_size;
+    return sum;
 }
-uint16_t SoundBuffer::average_pwr()
+
+int SoundBuffer::average_pwr()
 {
-    uint32_t sum=0;
-    for(uint16_t i=0;i<m_size;i++)
+    int sum = 0;
+    for(uint16_t i = 0; i < m_size; i++)
     {
         if(m_Buffer[i] < 0)
-            sum+=-1*m_Buffer[i];
+            sum += -1*m_Buffer[i];
         else
-            sum+=m_Buffer[i];
+            sum += m_Buffer[i];
     }
-    sum/=m_size;
-    return (uint16_t)sum;
+    sum /= m_size;
+    return sum;
 }
-bool SoundBuffer::write(uint16_t pos,int16_t value)
+
+bool SoundBuffer::write(int pos,int16_t value)
 {
         if(pos < m_size)
         {
-            m_Buffer[pos]=value;
+            m_Buffer[pos] = value;
             return true;
         }
         else
             return false;
 }
-int16_t SoundBuffer::read(uint16_t pos)
+
+int16_t SoundBuffer::read(int pos)
 {
         return m_Buffer.at(pos);
 }
-int16_t SoundBuffer::size()
+
+int SoundBuffer::size()
 {
     return m_size;
 }
+
+
 } //namespace libbeat
