+++ libbeat/soundbuffer.h
@@ -32,34 +32,35 @@ public:
     /*!
         @param m_size how many samples SoundBuffer should store
     */
-    SoundBuffer(uint16_t m_size=0);
+    SoundBuffer(int m_size = 0);
     /*! returns the average value of the SoundBuffer
         @return average value
     */
-    int16_t average();
+    int average();
     /*! returns the average power of the SoundBuffer (e.g the volume)
         @return average power
     */
-    uint16_t average_pwr();
+    int average_pwr();
     /*!
         write value to pos
         @param value the value
         @param pos the position
         @return true if successful, false if not
     */
-    bool write(uint16_t pos, int16_t value);
+    bool write(int pos, int16_t value);
     /*!
         get a certain positon
         @param pos the position
         @return the value at pos
     */
-    int16_t read(uint16_t pos);
-    int16_t size();
+    int16_t read(int pos);
+    int size();
 
 private:
-    uint16_t m_size;
+    int m_size;
     QVector<int16_t> m_Buffer;
 
 };
+
 }
 #endif // SOUNDBUFFER_H
